requirements.txt
keras==3.5.0
numpy==2.2.1
opencv_python==4.10.0.84
Pillow==11.0.0
scikit_learn==1.6.0
streamlit==1.37.0
